+++ src/rt/worker.c
@@ -190,8 +190,8 @@ do_pixel(int cpu, int pat_num, int pixelnum)
 	a.a_x <<= (incr_nlevel-incr_level);
 	a.a_y <<= (incr_nlevel-incr_level);
     } else {
-	a.a_y = pixelnum/width;
-	a.a_x = pixelnum - (a.a_y * width);
+	a.a_y = (int)(pixelnum/width);
+	a.a_x = (int)(pixelnum - (a.a_y * width));
 	/* a.a_x = pixelnum%width; */
     }
 
@@ -698,7 +698,7 @@ grid_setup(void)
 void
 do_run(int a, int b)
 {
-    int cpu;
+    size_t cpu;
 
 #ifdef USE_FORKED_THREADS
     int pid, wpid;
